Revert of commit 6d8f709fd8a5a3f16f02f53e0586bd8285700163.
This check was needed because the content_type may not exist for every sender.

+++ website/jdpages/signals.py
@@ -39,7 +39,8 @@ def pre_delete_callback(sender, instance, **kwargs):
     sender -- the model class
     instance -- the actual instance being saved
     """
-    related_elements = ColumnElement.objects.filter(object_id=instance.id, content_type=ContentType.objects.get_for_model(sender))
-    for element in related_elements:
-        element.delete()
+    if sender == BlogCategory:
+        related_elements = ColumnElement.objects.filter(object_id=instance.id, content_type=ContentType.objects.get_for_model(sender))
+        for element in related_elements:
+            element.delete()
     return
